benchmarker/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
docker==7.1.0
fonttools==4.51.0
idna==3.7
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pika==1.3.2
pillow==10.3.0
psutil==5.9.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gitlab==4.4.0
requests==2.31.0
requests-toolbelt==1.0.0
sentry-sdk==1.45.0
six==1.16.0
urllib3==2.2.1
aio_pika
asyncio
